[FIX] Make a better use of stale action operations
Before this patch, as you can see in [the logs attached here][1], the bot ran 3 times.

The last one is untouched, but the other two are now merged into a single execution. This way we do less API calls and get it running faster. Also the logs will become less confusing.

[1]: https://github.com/OCA/web/pull/2037#pullrequestreview-764000869