Run dev status and license checks in separate jobs.
This has the downside of running more jobs, with several parallel
installation of the addons. On the other hand this helps in situations
where INCLUDE/EXCLUDE do not install everything, breaking the
checks with missing dependencies. It should also makes it easier to
surface issues if we make these checks non blocking.