[FIX] Make Python version explicit in pre-commit workflow
Recently (or recently enough that I've just begun noticing it), the
nodeenv dependency had troubles installing. Sometimes. It wasn't very
reliably failing.

This is related to <https://github.com/ekalinin/nodeenv/issues/328>.

By explicitly setting the Python version to something <=3.10, we can
safely pretend the problem doesn't exist.

Ideally the Python versions match up to the ones used in test.yml, but
they don't, and I fear it might be too intrusive to fix that in this
commit.

Signed-off-by: Carmen Bianca BAKKER <carmen@coopiteasy.be>