fix(oca-gen-addon-readme): reduce unneeded merge conflicts
As seen in https://github.com/OCA/oca-addons-repo-template/pull/190#issuecomment-1784076360, when somebody tried to merge 2 PRs for the same module, chances are that both imply different readme digests and the merge fails.

As seen in https://github.com/OCA/oca-addons-repo-template/pull/220#issuecomment-1786814130, this was already happening under some circumstances with `requirements.txt` files.

Here I added a `.gitattributes` file which specifies a `merge=union` driver for those files. [Git docs][1] explain that this driver will just keep the changes from all files, marking the merge as succeeded and not including merge markers. This is not ideal, but is OK for our current situation because:

1. Duplicate/conflicting lines in those files should not really harm a lot.
2. The bot will re-render the README upon merge anyways, fixing that conflict automatically.
3. We could define a custom merge driver, but `union` is built into Git, meaning less configurations required downstream.

Still, it doesn't seem to work in octopus merges. At least, if you do consecutive merges instead, all will work as expected.

At the same time, I moved the hook to be the last. This is because previous hooks can reformat the files used to obtain the digest.

Closes #220.

[1]: https://git-scm.com/docs/gitattributes.html#Documentation/gitattributes.txt-union